+++ lib/links/widgets/link-menu.dart
@@ -1,12 +1,8 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import '../../controller/controllers/editor-controller.dart';
-import '../../doc-tree/services/coordinates.service.dart';
 import '../../document/models/attributes/attributes.model.dart';
-import '../../document/services/nodes/node.utils.dart';
 import '../../editor/services/editor.service.dart';
 import '../../selection/services/selection.service.dart';
 import '../../shared/state/editor-state-receiver.dart';
@@ -16,7 +12,7 @@ import '../../styles/services/styles.service.dart';
 import '../../toolbar/widgets/dialogs/link-style-dialog.dart';
 import '../services/links.service.dart';
 
-// When tapping links this menu opens, displaying the url of the link, 
+// When tapping links this menu opens, displaying the url of the link,
 // and 3 buttons (edit link, remove link from text/url, and copy link to clipboard)
 // ignore: must_be_immutable
 class LinkMenu extends StatefulWidget with EditorStateReceiver {
@@ -43,41 +39,31 @@ class _LinkMenuState extends State<LinkMenu> {
   late final EditorService _editorService;
   late final SelectionService _selectionService;
   late final StylesService _stylesService;
-  late final CoordinatesService _coordinatesService;
   late final LinksService _linksService;
 
   @override
   void initState() {
     _editorService = EditorService(widget._state);
     _selectionService = SelectionService(widget._state);
-    _coordinatesService = CoordinatesService(widget._state);
     _stylesService = StylesService(widget._state);
-    _linksService =  LinksService(widget._state);
+    _linksService = LinksService(widget._state);
     super.initState();
   }
 
   @override
-  Widget build(BuildContext context) {
-    final offset = _getOffsetForLinkMenu();
-
-    return Positioned(
-      top: offset.dy,
-      left: offset.dx,
-      child: TooltipMenu(
+  Widget build(BuildContext context) => TooltipMenu(
         children: [
           _linkUrl(),
           _removeLinkBtn(),
           _editLinkBtn(),
           _copyLinkToClipboardBtn(),
         ],
-      ),
-    );
-  }
+      );
 
   Widget _removeLinkBtn() => Material(
         color: Colors.transparent,
         child: InkWell(
-          onTap: _linksService.removeLink,
+          onTap: _linksService.removeSelectionLink,
           child: Container(
             padding: EdgeInsets.all(6),
             child: Icon(
@@ -139,53 +125,6 @@ class _LinkMenuState extends State<LinkMenu> {
       .attributes[AttributesM.link.key]
       ?.value;
 
-  Offset _getOffsetForLinkMenu() {
-    // Default value, if not set.
-    var offset = Offset.infinite;
-
-    final linkRect = widget._state.selectedLink.selectedLinkRectangles;
-    final hasRect = linkRect.isNotEmpty;
-
-    if (hasRect) {
-      final firstLinkRectIsNotEmpty = linkRect[0].rectangles.isNotEmpty;
-
-      if (firstLinkRectIsNotEmpty) {
-        final isLinkSelected =
-            _stylesService.getSelectionStyle().attributes.containsKey('link');
-
-        if (isLinkSelected) {
-          // Get positions for offset
-          final rectangle = linkRect[0].rectangles[0];
-          final lineOffset = linkRect[0].docRelPosition;
-
-          final childAtOffset = _coordinatesService.childAtOffset(lineOffset);
-          final docOffset = NodeUtils().getDocumentOffset(
-            childAtOffset.container,
-          );
-          final extentOffset = widget.controller.selection.extentOffset;
-          final lineHeight = childAtOffset.preferredLineHeight(
-            TextPosition(
-              offset: extentOffset - docOffset,
-            ),
-          );
-          final hLeftPoint = rectangle.left;
-          const triangleTooltipMargin = 10;
-
-          // Menu Position
-          offset = Offset(
-            hLeftPoint - triangleTooltipMargin,
-            lineOffset.dy + lineHeight + rectangle.top,
-          );
-
-          return offset;
-        }
-      }
-    }
-
-    // Fail case
-    return offset;
-  }
-
   void _openLinkDialog(BuildContext context) {
     showDialog<dynamic>(
       context: context,
@@ -208,8 +147,10 @@ class _LinkMenuState extends State<LinkMenu> {
         final len = selection.end - index;
         text ??= len == 0
             ? ''
-            : widget._state.refs.documentController
-            .getPlainTextAtRange(index, len);
+            : widget._state.refs.documentController.getPlainTextAtRange(
+                index,
+                len,
+              );
 
         return LinkStyleDialog(
           link: link,
@@ -217,7 +158,7 @@ class _LinkMenuState extends State<LinkMenu> {
         );
       },
     ).then(
-          (value) {
+      (value) {
         if (value != null) _editorService.addLinkToSelection(value);
       },
     );