+++ lib/selection/state/selected-link.state.dart
@@ -4,9 +4,9 @@ class SelectedLinkState {
   // === SELECTED LINK RECTANGLES ===
 
   // After selecting a link, we need to cache its rectangles in order to display the link menu
-  List<SelectionRectanglesM> _selectedLinkRectangles = [];
+  late List<SelectionRectanglesM> _selectedLinkRectangles;
 
-  List<SelectionRectanglesM> get selectedLinkRectangles => _selectedLinkRectangles;
+  List<SelectionRectanglesM>? get selectedLinkRectangles => _selectedLinkRectangles;
 
   void setSelectedLinkRectangles(List<SelectionRectanglesM> rectangles) =>
       _selectedLinkRectangles = rectangles;